Rename feature-link.js to chromedash-link.js.
